Move combine_lookups into serializatio/constraints
