+++ msm/src/lookups.rs
@@ -1,15 +1,8 @@
 //! Instantiate the MVLookup protocol for the MSM project.
 
-use crate::{
-    columns::Column,
-    expr::{curr_cell, E},
-    mvlookup::{LookupTableID, MVLookup, MVLookupWitness},
-};
-use ark_ff::{FftField, Field, Zero};
-use kimchi::circuits::{
-    domains::EvaluationDomains,
-    expr::{ChallengeTerm, ConstantExpr, ExprInner},
-};
+use crate::mvlookup::{LookupTableID, MVLookup, MVLookupWitness};
+use ark_ff::{FftField, Field};
+use kimchi::circuits::domains::EvaluationDomains;
 use rand::{seq::SliceRandom, thread_rng, Rng};
 use std::iter;
 
@@ -116,73 +109,4 @@ impl<F: FftField> LookupWitness<F> {
             m,
         }
     }
-}
-
-/// Compute the following constraint:
-/// ```text
-///                     lhs
-///    |------------------------------------------|
-///    |                           denominators   |
-///    |                         /--------------\ |
-/// column * (\prod_{i = 1}^{N} (\beta + f_{i}(X))) =
-/// \sum_{i = 1}^{N} m_{i} * \prod_{j = 1, j \neq i}^{N} (\beta + f_{j}(X))
-///    |             |--------------------------------------------------|
-///    |                             Inner part of rhs                  |
-///    |                                                                |
-///    |                                                               /
-///     \                                                             /
-///      \                                                           /
-///       \---------------------------------------------------------/
-///                           rhs
-/// ```
-pub fn combine_lookups<F: Field>(column: Column, lookups: Vec<Lookup<E<F>>>) -> E<F> {
-    let joint_combiner = {
-        let joint_combiner = ConstantExpr::from(ChallengeTerm::JointCombiner);
-        E::Atom(ExprInner::Constant(joint_combiner))
-    };
-    let beta = {
-        let beta = ConstantExpr::from(ChallengeTerm::Beta);
-        E::Atom(ExprInner::Constant(beta))
-    };
-
-    // Compute (\beta + f_{i}(X)) for each i.
-    // Note that f_i(X) = x_{0} + r x_{1} + ... r^{N} x_{N} + r^{N + 1} table_id
-    let denominators = lookups
-        .iter()
-        .map(|x| {
-            let combined_value = (x
-                .value
-                .iter()
-                .rev()
-                .fold(E::zero(), |acc, y| acc * joint_combiner.clone() + y.clone())
-                * joint_combiner.clone())
-                + x.table_id.to_constraint();
-            beta.clone() + combined_value
-        })
-        .collect::<Vec<_>>();
-    // Compute `column * (\prod_{i = 1}^{N} (\beta + f_{i}(X)))`
-    let lhs = denominators
-        .iter()
-        .fold(curr_cell(column), |acc, x| acc * x.clone());
-    let rhs = lookups
-        .into_iter()
-        .enumerate()
-        .map(|(i, x)| {
-            denominators.iter().enumerate().fold(
-                // Compute individual \sum_{j = 1, j \neq i}^{N} f_{j}(X)
-                // This is the inner part of rhs. It multiplies with m_{i}
-                x.numerator,
-                |acc, (j, y)| {
-                    if i == j {
-                        acc
-                    } else {
-                        acc * y.clone()
-                    }
-                },
-            )
-        })
-        // Individual sums
-        .reduce(|x, y| x + y)
-        .unwrap_or(E::zero());
-    lhs - rhs
 }
