+++ msm/src/serialization/constraints.rs
@@ -1,13 +1,87 @@
-use ark_ff::PrimeField;
+use ark_ff::{Field, PrimeField, Zero};
 use kimchi::circuits::{
-    expr::{ConstantExpr, ConstantTerm, Expr, ExprInner, Variable},
+    expr::{ChallengeTerm, ConstantExpr, ConstantTerm, Expr, ExprInner, Variable},
     gate::CurrOrNext,
 };
 
-use crate::{columns::Column, expr::E, serialization::N_INTERMEDIATE_LIMBS, N_LIMBS};
+use crate::{
+    columns::Column,
+    expr::{curr_cell, E},
+    serialization::N_INTERMEDIATE_LIMBS,
+    MVLookupTableID as _, N_LIMBS,
+};
 
 use super::Lookup;
 use super::{interpreter::InterpreterEnv, LookupTable};
+
+/// Compute the following constraint:
+/// ```text
+///                     lhs
+///    |------------------------------------------|
+///    |                           denominators   |
+///    |                         /--------------\ |
+/// column * (\prod_{i = 1}^{N} (\beta + f_{i}(X))) =
+/// \sum_{i = 1}^{N} m_{i} * \prod_{j = 1, j \neq i}^{N} (\beta + f_{j}(X))
+///    |             |--------------------------------------------------|
+///    |                             Inner part of rhs                  |
+///    |                                                                |
+///    |                                                               /
+///     \                                                             /
+///      \                                                           /
+///       \---------------------------------------------------------/
+///                           rhs
+/// ```
+pub fn combine_lookups<F: Field>(column: Column, lookups: Vec<Lookup<E<F>>>) -> E<F> {
+    let joint_combiner = {
+        let joint_combiner = ConstantExpr::from(ChallengeTerm::JointCombiner);
+        E::Atom(ExprInner::Constant(joint_combiner))
+    };
+    let beta = {
+        let beta = ConstantExpr::from(ChallengeTerm::Beta);
+        E::Atom(ExprInner::Constant(beta))
+    };
+
+    // Compute (\beta + f_{i}(X)) for each i.
+    // Note that f_i(X) = x_{0} + r x_{1} + ... r^{N} x_{N} + r^{N + 1} table_id
+    let denominators = lookups
+        .iter()
+        .map(|x| {
+            let combined_value = (x
+                .value
+                .iter()
+                .rev()
+                .fold(E::zero(), |acc, y| acc * joint_combiner.clone() + y.clone())
+                * joint_combiner.clone())
+                + x.table_id.to_constraint();
+            beta.clone() + combined_value
+        })
+        .collect::<Vec<_>>();
+    // Compute `column * (\prod_{i = 1}^{N} (\beta + f_{i}(X)))`
+    let lhs = denominators
+        .iter()
+        .fold(curr_cell(column), |acc, x| acc * x.clone());
+    let rhs = lookups
+        .into_iter()
+        .enumerate()
+        .map(|(i, x)| {
+            denominators.iter().enumerate().fold(
+                // Compute individual \sum_{j = 1, j \neq i}^{N} f_{j}(X)
+                // This is the inner part of rhs. It multiplies with m_{i}
+                x.numerator,
+                |acc, (j, y)| {
+                    if i == j {
+                        acc
+                    } else {
+                        acc * y.clone()
+                    }
+                },
+            )
+        })
+        // Individual sums
+        .reduce(|x, y| x + y)
+        .unwrap_or(E::zero());
+    lhs - rhs
+}
 
 pub struct Env<Fp> {
     /// An indexed set of constraints.
@@ -32,6 +106,14 @@ impl<Fp: PrimeField> Env<Fp> {
 }
 
 impl<F: PrimeField> InterpreterEnv<F> for Env<F> {
+    /// Schemas for the columns:
+    /// - kimchi_limbs: 0 to 2
+    /// - msm_limbs: 3 to 20
+    /// - intermediate_limbs: 21 to 40
+    /// - lookup_aggreg: 41
+    /// - lookup multiplicities: 42 and 43 (2 tables)
+    /// - 3 lookup partial sums RC15: 44 to 46
+    /// - 4 lookup partial sums RC4: 48 to 51
     type Position = Column;
 
     type Variable = E<F>;
